Some refactor
