--- src/core/default/PageModel.php
@@ -1,11 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: sergiovilar
- * Date: 12/07/14
- * Time: 12:09 AM
- */
-
-class PageModel extends AppModel{
-
-}
