--- src/core/default/PostModel.php
@@ -1,11 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: sergiovilar
- * Date: 12/07/14
- * Time: 12:08 AM
- */
-
-class PostModel extends AppModel{
-
-}
